Reorg: Move asset to scene instead of util
