+++ core/src/scene/asset.h
